Add threading support in CMakeLists for tests
The CMakeLists.txt file for the tests module has been updated to include Threads as a required package. Each test is now built as an executable that is linked with the Threads library. This is done by iterating over all source files, then for each source file, an executable is created and linked with the Threads library. Moreover, tests are also added to further utilize the updated setup.